Include entity_type from KG
This is so we can ensure uniqueness between entities, their entity type and the synonyms generated.

We include this because we can have an entity with more than one entity_type. For instance, JSA can be a Finance and State entity_type.

Improve naming of cleaned processed content store file and check for uniqueness in synonyms data.